Changed how we do the bootstrapping
Now we take random samples from the entire population instead of just resampling the RES pop to make a new RES dataset and the SUS pop for a SUS dataset.
To get the p-value we then determine how often the diff between the simulated RES diversity and the simulated SUS diversity is bigger than or equal to the real RES diversity and the real SUS diversity.